refactor: implement PR code feedback
